Update publishing component labeling, tooltip and badge
